Fixed typecasting error for float()

+++ source/data_read_helper.py
@@ -231,7 +231,7 @@ def check_valid_chargeamount(data_amount):
     '''
         check if charge amount satisfies the condition
     '''
-    if int(data_amount) >= 0:
+    if float(data_amount) >= 0:
         return True
 
     return False
@@ -240,7 +240,7 @@ def check_valid_priceamount(data_amount):
     '''
         check if the price is positive and return bool
     '''
-    if int(data_amount) >= 0:
+    if float(data_amount) >= 0:
         return True
 
     return False
